backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.2
beautifulsoup4==4.12.2
pydantic>=2.5.3
python-multipart==0.0.6
python-dotenv==1.0.0
xai-sdk==1.3.1